requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
numpy==1.26.2 